Add files via upload
Welcome to my personal portfolio website! This project is the culmination of my skills and experiences, providing a showcase of my projects, skills, and background. Built with HTML and CSS, this website is designed to introduce visitors to who I am, what I do, and the projects I've worked on.